Address PR feedback: Remove files that don't belong in source tree and improve code quality
- Remove IMPLEMENTATION_SUMMARY.md, example_usage.py, and examples.py per feedback
- Remove redundant property comments in enums.py that violate commenting guidelines
- Replace sample data in production views.py with proper UI testing pattern
- Add UI testing infrastructure in tests/ui/ directory with synthetic data
- Remove confusing backwards compatibility comments from new feature
- Update tests to reflect Phase 1 infrastructure-only implementation
- All tests passing, linting clean

Addresses feedback from anthony-rocco in PR review
