+++ src/hi/apps/monitor/templates/monitor/tests/ui/home.html
@@ -1,0 +1,52 @@
+{% extends "pages/base.html" %}
+{% load icons %}
+
+{% block head_title %}HI: Monitor UI Tests{% endblock %}
+
+{% block content %}
+<div class="container-fluid m-4">
+  <h2 class="text-info">Monitor UI Tests</h2>
+
+  <div class="row">
+    <div class="col-md-6">
+      <h3>Monitor Health Status Modals</h3>
+      <ul class="list-group">
+        <li class="list-group-item">
+          <a href="{% url 'test_monitor_health_status' monitor_type='zoneminder' %}" data-toggle="modal" data-target="#monitorHealthModal">
+            ZoneMinder - Healthy Single API
+          </a>
+        </li>
+        <li class="list-group-item">
+          <a href="{% url 'test_monitor_health_status' monitor_type='weather' %}" data-toggle="modal" data-target="#monitorHealthModal">
+            Weather - Warning Multiple APIs
+          </a>
+        </li>
+        <li class="list-group-item">
+          <a href="{% url 'test_monitor_health_status' monitor_type='alert' %}" data-toggle="modal" data-target="#monitorHealthModal">
+            Alert - Error No APIs
+          </a>
+        </li>
+      </ul>
+    </div>
+
+    <div class="col-md-6">
+      <h3>Instructions</h3>
+      <p>Click the links to test different monitor health status scenarios:</p>
+      <ul>
+        <li><strong>ZoneMinder:</strong> Single API source, healthy status</li>
+        <li><strong>Weather:</strong> Multiple API sources, mixed health states</li>
+        <li><strong>Alert:</strong> No API sources, error condition</li>
+      </ul>
+    </div>
+  </div>
+</div>
+
+<!-- Modal placeholder for testing -->
+<div class="modal fade" id="monitorHealthModal" tabindex="-1" role="dialog" aria-labelledby="monitorHealthModalLabel" aria-hidden="true">
+  <div class="modal-dialog modal-lg" role="document">
+    <div class="modal-content">
+      <!-- Content loaded via AJAX -->
+    </div>
+  </div>
+</div>
+{% endblock %}
